server/requirements.txt
python-socketio==5.6
aiohttp==xx